Refactoring event popup handling
