+++ src/components/map/EventPopup.js
@@ -1,82 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import i18n from '@dhis2/d2-i18n';
-import { getInstance as getD2 } from 'd2';
 import Popup from './Popup';
 import { apiFetch } from '../../util/api';
-import {
-    getDisplayPropertyUrl,
-    formatTime,
-    formatCoordinate,
-} from '../../util/helpers';
+import { formatTime, formatCoordinate } from '../../util/helpers';
 
 // Returns true if value is not undefined or null;
 const hasValue = value => value !== undefined || value !== null;
-
-// Loads an option set for an data element to get option names
-const loadOptionSet = async dataElement => {
-    const { optionSet } = dataElement;
-
-    if (!optionSet || !optionSet.id) {
-        return dataElement;
-    }
-
-    const d2 = await getD2();
-
-    if (optionSet && optionSet.id) {
-        const fullOptionSet = await d2.models.optionSets.get(optionSet.id, {
-            fields:
-                'id,displayName~rename(name),options[code,displayName~rename(name)]',
-            paging: false,
-        });
-
-        if (fullOptionSet && fullOptionSet.options) {
-            dataElement.options = fullOptionSet.options.reduce(
-                (byId, option) => {
-                    byId[option.code] = option.name;
-                    return byId;
-                },
-                {}
-            );
-        }
-    }
-};
-
-// Loads the data elements for a program stage to display in popup
-const loadDataElements = async (programStage, eventCoordinateField) => {
-    const d2 = await getD2();
-    const data = await d2.models.programStage.get(programStage.id, {
-        fields: `programStageDataElements[displayInReports,dataElement[id,${getDisplayPropertyUrl(
-            d2
-        )},optionSet,valueType]]`,
-        paging: false,
-    });
-    const { programStageDataElements } = data;
-    let displayElements = [];
-    let eventCoordinateFieldName;
-
-    if (Array.isArray(programStageDataElements)) {
-        displayElements = programStageDataElements
-            .filter(d => d.displayInReports)
-            .map(d => d.dataElement);
-
-        for (let d of displayElements) {
-            await loadOptionSet(d);
-        }
-
-        if (eventCoordinateField) {
-            const coordElement = programStageDataElements.find(
-                d => d.dataElement.id === eventCoordinateField
-            );
-
-            if (coordElement) {
-                eventCoordinateFieldName = coordElement.dataElement.name;
-            }
-        }
-    }
-
-    return { displayElements, eventCoordinateFieldName };
-};
 
 // Loads event data for the selected feature
 const loadEventData = async feature =>
@@ -132,30 +62,13 @@ const EventPopup = props => {
     const {
         coordinates,
         feature,
-        programStage,
-        eventCoordinateField,
         styleDataItem,
+        displayElements,
+        eventCoordinateFieldName,
         onClose,
     } = props;
 
     const [eventData, setEventData] = useState();
-    const [dataElements, setDataElements] = useState({});
-    const { displayElements = [], eventCoordinateFieldName } = dataElements;
-
-    // Load data elements every time programStage is changed
-    useEffect(() => {
-        let aborted = false;
-
-        loadDataElements(programStage, eventCoordinateField).then(data => {
-            if (!aborted) {
-                setDataElements(data);
-            }
-        });
-
-        return () => {
-            aborted = true;
-        };
-    }, [programStage, eventCoordinateField, setDataElements]);
 
     // Load event data every time a new feature is clicked
     useEffect(() => {
@@ -168,11 +81,12 @@ const EventPopup = props => {
         });
 
         return () => {
+            setEventData(); // Clear event data
             aborted = true;
         };
     }, [feature, setEventData]);
 
-    if (!coordinates || !feature || !eventData) {
+    if (!eventData) {
         return null;
     }
 
@@ -183,10 +97,7 @@ const EventPopup = props => {
     return (
         <Popup
             coordinates={coordinates}
-            onClose={() => {
-                onClose();
-                setEventData();
-            }}
+            onClose={onClose}
             className="dhis2-map-popup-event"
         >
             <table>
@@ -208,14 +119,18 @@ const EventPopup = props => {
                             </td>
                         </tr>
                     )}
-                    <tr>
-                        <th>{i18n.t('Organisation unit')}</th>
-                        <td>{orgUnitName}</td>
-                    </tr>
-                    <tr>
-                        <th>{i18n.t('Event time')}</th>
-                        <td>{formatTime(eventDate)}</td>
-                    </tr>
+                    {orgUnitName && (
+                        <tr>
+                            <th>{i18n.t('Organisation unit')}</th>
+                            <td>{orgUnitName}</td>
+                        </tr>
+                    )}
+                    {eventDate && (
+                        <tr>
+                            <th>{i18n.t('Event time')}</th>
+                            <td>{formatTime(eventDate)}</td>
+                        </tr>
+                    )}
                 </tbody>
             </table>
         </Popup>
@@ -223,10 +138,10 @@ const EventPopup = props => {
 };
 
 EventPopup.propTypes = {
-    coordinates: PropTypes.array,
-    feature: PropTypes.object,
-    programStage: PropTypes.object,
-    eventCoordinateField: PropTypes.string,
+    coordinates: PropTypes.array.isRequired,
+    feature: PropTypes.object.isRequired,
+    displayElements: PropTypes.array.isRequired,
+    eventCoordinateFieldName: PropTypes.string,
     styleDataItem: PropTypes.object,
     onClose: PropTypes.func.isRequired,
 };
